requirements.txt
fuzzywuzzy[speedup]
celery
beautifulsoup4
python-slugify
scrapy
colorama
selenium
pathvalidate
Pillow
binaryornot
tldextract
requests-cache
furl
jsonpath-rw
jsonpath-rw-ext
ImageHash
boto3
PrettyTable